+++ CHANGELOG.md
@@ -7,12 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `amountType`, `paymentAmount` and `paymentCalculation` to payment event amount configuration
+
 ## [13.2.2] - 2026-07-28
 
 ### Added
 
 - `SUBMISSION_TIMESTAMP` type to workflow event and approval step conditional predicates schema
-- `amountType`, `paymentAmount` and `paymentCalculation` to payment event amount configuration
 
 ### Dependencies