+++ CHANGELOG.md
@@ -7,11 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [13.2.3] - 2026-08-26
+
 ### Added
 
 - `amountType`, `paymentAmount` and `paymentCalculation` to payment event amount configuration
 - `screenReaderAlert` to html element validation
 - `hiddenFrom` on form elements so they can be hidden from submitters, approvers, or both
+
+### Dependencies
+
+- update @oneblink/sdk-core to [10.1.0-beta.9](https://github.com/oneblink/sdk-core-js/blob/master/CHANGELOG.md) (from [10.0.0-beta.3](https://github.com/oneblink/sdk-core-js/blob/master/CHANGELOG.md))
+
+- update @oneblink/storage to [8.0.0-beta.2](https://github.com/oneblink/storage/blob/master/CHANGELOG.md) (from [7.1.2-beta.2](https://github.com/oneblink/storage/blob/master/CHANGELOG.md))
 
 ## [13.2.2] - 2026-07-28